LayoutTests/accessibility/mac/frame-with-title-expected.txt


This test makes sure that a frame with a title will use that value instead of the name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.description is 'AXDescription: All Packages'
PASS accessibilityController.focusedElement.description is 'AXDescription: packageListFrame'
PASS successfullyParsed is true

TEST COMPLETE
